extracted the validation code into its component
