+++ client/src/component/SignupForm.js
@@ -2,35 +2,14 @@ import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import "./SignupForm.css";
-
+import {formValidation} from "./statics/FormValidation"
 import content from "./statics/InputsData";
-import * as yup from "yup";
+
 
 export default function SignupForm() {
-  console.log(useForm());
 
-  let schema = yup.object().shape({
-    firstName: yup.string().required("First name is required!"),
-    surname: yup.string().required("Surname is required!"),
-    email: yup
-      .string()
-      .required("Email is required !")
-      .email("Email is invalid"),
-    confirmEmail: yup
-      .string()
-      .oneOf([yup.ref("email"), null], "email must match")
-      .required("Field is required please fill it !"),
-    password: yup
-      .string()
-      .required("Password is required")
-      .min(6, "Password must be at least 6 characters"),
-    confirmPassword: yup
-      .string()
-      .oneOf([yup.ref("password"), null], "Passwords must match")
-      .required("Confirm Password is required"),
-    city: yup.string().required("City is required!"),
-    phoneNumber: yup.number().required().positive(),
-  });
+let schema = formValidation()
+
   const { register, handleSubmit, errors, reset } = useForm({
     resolver: yupResolver(schema),
   });
